requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.30.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.2.0
pytest-cov==4.1.0 